finalized refactoring names
